fix reset and double

+++ vowpalwabbit/automl.h
@@ -109,7 +109,7 @@ struct interaction_config_manager : config_manager
   std::string oracle_type;
   dense_parameters& weights;
   priority_func* calc_priority;
-  float confidence_interval;
+  double confidence_interval;
 
   // Stores all namespaces currently seen -- Namespace switch could we use array, ask Jack
   std::map<namespace_index, uint64_t> ns_counter;
@@ -124,7 +124,7 @@ struct interaction_config_manager : config_manager
   std::priority_queue<std::pair<float, uint64_t>> index_queue;
 
   interaction_config_manager(uint64_t, uint64_t, std::shared_ptr<VW::rand_state>, uint64_t, bool, std::string,
-      dense_parameters&, float (*)(const exclusion_config&, const std::map<namespace_index, uint64_t>&), float);
+      dense_parameters&, float (*)(const exclusion_config&, const std::map<namespace_index, uint64_t>&), double);
 
   void apply_config(example*, uint64_t);
   void revert_config(example*);
